+++ frontend/README.md
@@ -1,0 +1,65 @@
+# VibeCheck Frontend
+
+This is the frontend application for VibeCheck, built with React and Vite.
+
+## Prerequisites
+
+- Node.js (v16 or higher)
+- npm or yarn
+
+## Installation
+
+Install the dependencies:
+
+```bash
+npm install
+```
+
+## Development Server
+
+To start the development server:
+
+```bash
+npm run dev
+```
+
+The application will be available at `http://localhost:5173` (default Vite port).
+
+## Build for Production
+
+To create a production build:
+
+```bash
+npm run build
+```
+
+The built files will be output to the `dist/` directory.
+
+## Preview Production Build
+
+To preview the production build locally:
+
+```bash
+npm run preview
+```
+
+## Linting
+
+To run ESLint:
+
+```bash
+npm run lint
+```
+
+## Project Structure
+
+- `src/` - Source files
+  - `App.jsx` - Main application component
+  - `Login.jsx` - Login page component
+  - `Register.jsx` - Registration page component
+  - `main.jsx` - Application entry point
+  - `assets/` - Static assets
+
+## Configuration
+
+The Vite configuration can be found in `vite.config.js`.
