--- frontend/frontend/.env.example
@@ -1,1 +1,0 @@
-VITE_API_BASE_URL=http://localhost:3000/api/v1
