--- frontend/frontend/src/Profile.css
@@ -1,412 +1,0 @@
-.profile-container {
-  max-width: 900px;
-  margin: 0 auto;
-  padding: 2rem;
-}
-
-.profile-section {
-  background: linear-gradient(135deg, #f5f7fa 0%, #c3cfe2 100%);
-  border-radius: 12px;
-  padding: 2rem;
-  margin-bottom: 2rem;
-  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
-}
-
-.section-header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin-bottom: 2rem;
-  border-bottom: 2px solid rgba(0, 0, 0, 0.1);
-  padding-bottom: 1rem;
-}
-
-.section-header h2 {
-  margin: 0;
-  color: #2c3e50;
-  font-size: 1.5rem;
-}
-
-.edit-btn,
-.save-btn,
-.add-btn {
-  padding: 0.75rem 1.5rem;
-  border: none;
-  border-radius: 8px;
-  background: #667eea;
-  color: white;
-  cursor: pointer;
-  font-weight: 600;
-  transition: all 0.3s ease;
-}
-
-.edit-btn:hover,
-.save-btn:hover,
-.add-btn:hover {
-  background: #5568d3;
-  transform: translateY(-2px);
-  box-shadow: 0 4px 12px rgba(102, 126, 234, 0.4);
-}
-
-.profile-content {
-  display: flex;
-  flex-direction: column;
-  gap: 1.5rem;
-}
-
-.profile-field {
-  display: flex;
-  flex-direction: column;
-  gap: 0.5rem;
-}
-
-.profile-field label {
-  font-weight: 600;
-  color: #2c3e50;
-  font-size: 0.95rem;
-}
-
-.field-value {
-  color: #555;
-  font-size: 1rem;
-  margin: 0;
-  padding: 0.75rem;
-  background: rgba(255, 255, 255, 0.7);
-  border-radius: 6px;
-  word-break: break-word;
-}
-
-.text-input,
-.textarea-input,
-.color-input {
-  padding: 0.75rem;
-  border: 2px solid #e0e0e0;
-  border-radius: 6px;
-  font-size: 1rem;
-  font-family: inherit;
-  transition: all 0.3s ease;
-}
-
-.text-input:focus,
-.textarea-input:focus,
-.color-input:focus {
-  outline: none;
-  border-color: #667eea;
-  box-shadow: 0 0 0 3px rgba(102, 126, 234, 0.1);
-}
-
-.textarea-input {
-  resize: vertical;
-  min-height: 100px;
-}
-
-.avatar-preview {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  height: 150px;
-  background: white;
-  border-radius: 8px;
-  overflow: hidden;
-}
-
-.avatar-preview img {
-  max-width: 100%;
-  max-height: 100%;
-  object-fit: cover;
-}
-
-/* Tags Editor */
-.tags-editor {
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-}
-
-.tag-input-group {
-  display: flex;
-  gap: 0.5rem;
-}
-
-.tag-input-group input {
-  flex: 1;
-}
-
-.tags-display,
-.colors-display {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 0.75rem;
-}
-
-.tag {
-  display: inline-flex;
-  align-items: center;
-  gap: 0.5rem;
-  padding: 0.5rem 1rem;
-  background: #667eea;
-  color: white;
-  border-radius: 20px;
-  font-size: 0.9rem;
-  font-weight: 500;
-}
-
-.remove-tag-btn,
-.remove-color-btn {
-  background: none;
-  border: none;
-  color: inherit;
-  cursor: pointer;
-  font-size: 1.2rem;
-  padding: 0;
-  line-height: 1;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 20px;
-  height: 20px;
-  transition: all 0.2s ease;
-}
-
-.remove-tag-btn:hover,
-.remove-color-btn:hover {
-  background: rgba(0, 0, 0, 0.2);
-  border-radius: 50%;
-}
-
-/* Colors Editor */
-.colors-editor {
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-}
-
-.color-input-group {
-  display: flex;
-  gap: 0.5rem;
-  align-items: center;
-}
-
-.color-input {
-  width: 60px;
-  height: 40px;
-  cursor: pointer;
-  padding: 4px;
-}
-
-.color-item {
-  display: flex;
-  align-items: center;
-  gap: 0.75rem;
-  padding: 0.75rem 1rem;
-  background: white;
-  border-radius: 8px;
-  border: 1px solid #e0e0e0;
-}
-
-.color-swatch {
-  width: 30px;
-  height: 30px;
-  border-radius: 4px;
-  border: 1px solid rgba(0, 0, 0, 0.1);
-  flex-shrink: 0;
-}
-
-.color-item span {
-  font-family: monospace;
-  font-size: 0.9rem;
-  color: #555;
-  flex: 1;
-}
-
-/* Categories Stats */
-.categories-stats {
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-}
-
-.category-stat {
-  display: grid;
-  grid-template-columns: 80px 1fr 50px;
-  align-items: center;
-  gap: 1rem;
-}
-
-.category-name {
-  font-weight: 600;
-  color: #2c3e50;
-  text-transform: capitalize;
-}
-
-.progress-bar {
-  background: #e0e0e0;
-  border-radius: 10px;
-  height: 8px;
-  overflow: hidden;
-}
-
-.progress-fill {
-  background: linear-gradient(90deg, #667eea 0%, #764ba2 100%);
-  height: 100%;
-  transition: width 0.3s ease;
-}
-
-.percentage {
-  font-weight: 600;
-  color: #667eea;
-  text-align: right;
-}
-
-/* Shares Section */
-.shares-grid {
-  display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(200px, 1fr));
-  gap: 1.5rem;
-  margin-top: 1rem;
-}
-
-.share-card {
-  background: white;
-  border-radius: 12px;
-  overflow: hidden;
-  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.1);
-  transition: all 0.3s ease;
-  border: 1px solid #e0e0e0;
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-}
-
-.share-card:hover {
-  transform: translateY(-4px);
-  box-shadow: 0 8px 20px rgba(0, 0, 0, 0.15);
-}
-
-.share-image-container {
-  position: relative;
-  width: 100%;
-  height: 180px;
-  background: linear-gradient(135deg, #f5f7fa 0%, #c3cfe2 100%);
-  overflow: hidden;
-}
-
-.share-image {
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-  display: block;
-}
-
-.dominant-color-indicator {
-  position: absolute;
-  top: 8px;
-  right: 8px;
-  width: 28px;
-  height: 28px;
-  border-radius: 6px;
-  border: 2px solid white;
-  box-shadow: 0 2px 6px rgba(0, 0, 0, 0.25);
-  cursor: help;
-}
-
-.share-category-badge {
-  position: absolute;
-  bottom: 8px;
-  left: 8px;
-  display: flex;
-  align-items: center;
-  gap: 0.4rem;
-  padding: 0.4rem 0.8rem;
-  background: rgba(102, 126, 234, 0.9);
-  color: white;
-  border-radius: 6px;
-  font-size: 0.75rem;
-  font-weight: 600;
-  text-transform: capitalize;
-  backdrop-filter: blur(10px);
-}
-
-.share-category-badge svg {
-  flex-shrink: 0;
-}
-
-.share-info {
-  padding: 1rem;
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  gap: 0.5rem;
-}
-
-.share-title {
-  margin: 0;
-  color: #2c3e50;
-  font-size: 0.95rem;
-  font-weight: 600;
-  line-height: 1.3;
-  display: -webkit-box;
-  -webkit-line-clamp: 2;
-  -webkit-box-orient: vertical;
-  overflow: hidden;
-}
-
-.share-caption {
-  margin: 0.25rem 0 0 0;
-  color: #666;
-  font-size: 0.85rem;
-  line-height: 1.4;
-  display: -webkit-box;
-  -webkit-line-clamp: 2;
-  -webkit-box-orient: vertical;
-  overflow: hidden;
-}
-
-.share-date {
-  margin-top: auto;
-  margin-bottom: 0;
-  color: #999;
-  font-size: 0.75rem;
-  font-weight: 500;
-}
-
-.empty-state {
-  text-align: center;
-  color: #999;
-  padding: 2rem 1rem;
-  font-style: italic;
-}
-
-.error-message {
-  background: #fee;
-  border: 1px solid #fcc;
-  color: #c33;
-  padding: 1rem;
-  border-radius: 6px;
-  margin-bottom: 1.5rem;
-  font-weight: 500;
-}
-
-@media (max-width: 768px) {
-  .profile-container {
-    padding: 1rem;
-  }
-
-  .profile-section {
-    padding: 1.5rem;
-  }
-
-  .section-header {
-    flex-direction: column;
-    gap: 1rem;
-    align-items: flex-start;
-  }
-
-  .category-stat {
-    grid-template-columns: 1fr;
-  }
-
-  .shares-grid {
-    grid-template-columns: 1fr;
-  }
-}
