--- frontend/frontend/src/components/stories/StoryCustomizer.css
@@ -1,156 +1,0 @@
-.story-customizer {
-  display: flex;
-  flex-direction: column;
-  gap: 0;
-  animation: fadeSlideUp 0.3s ease both;
-}
-
-.story-customizer__toggle {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  padding: 12px 16px;
-  border-radius: 14px;
-  border: 1.5px solid rgba(29, 27, 34, 0.1);
-  background: rgba(255, 255, 255, 0.6);
-  font-size: 0.88rem;
-  font-weight: 650;
-  color: rgba(29, 27, 34, 0.7);
-  cursor: pointer;
-  transition: all 0.2s ease;
-  outline: none;
-  -webkit-tap-highlight-color: transparent;
-}
-
-.story-customizer__toggle:hover {
-  background: rgba(255, 255, 255, 0.9);
-  border-color: rgba(108, 99, 255, 0.3);
-}
-
-.story-customizer__body {
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-  padding: 16px;
-  margin-top: 8px;
-  border-radius: 16px;
-  background: rgba(255, 255, 255, 0.5);
-  border: 1px solid rgba(29, 27, 34, 0.08);
-  animation: fadeSlideUp 0.25s ease both;
-}
-
-.story-customizer__section {
-  display: flex;
-  flex-direction: column;
-  gap: 8px;
-}
-
-.story-customizer__section-label {
-  display: flex;
-  align-items: center;
-  gap: 6px;
-  font-size: 0.78rem;
-  font-weight: 650;
-  color: rgba(29, 27, 34, 0.5);
-  text-transform: uppercase;
-  letter-spacing: 0.06em;
-}
-
-.story-customizer__chips {
-  display: flex;
-  gap: 6px;
-  flex-wrap: wrap;
-}
-
-.story-customizer__chip {
-  display: inline-flex;
-  align-items: center;
-  gap: 5px;
-  padding: 7px 14px;
-  border-radius: 999px;
-  border: 1.5px solid rgba(29, 27, 34, 0.1);
-  background: rgba(255, 255, 255, 0.7);
-  font-size: 0.78rem;
-  font-weight: 600;
-  color: rgba(29, 27, 34, 0.65);
-  cursor: pointer;
-  transition: all 0.2s ease;
-  outline: none;
-  -webkit-tap-highlight-color: transparent;
-}
-
-.story-customizer__chip:hover {
-  background: rgba(255, 255, 255, 0.95);
-  border-color: rgba(108, 99, 255, 0.3);
-}
-
-.story-customizer__chip--active {
-  background: linear-gradient(135deg, #6b6bf8, #b44aff) !important;
-  color: #fff !important;
-  border-color: transparent;
-  box-shadow: 0 2px 10px rgba(107, 107, 248, 0.3);
-}
-
-.story-customizer__chip-preview {
-  font-size: 0.85rem;
-  font-weight: 700;
-  opacity: 0.8;
-}
-
-/* Font preview classes */
-.story-customizer__chip--font-default .story-customizer__chip-preview {
-  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", sans-serif;
-}
-.story-customizer__chip--font-serif .story-customizer__chip-preview {
-  font-family: "Georgia", "Times New Roman", serif;
-}
-.story-customizer__chip--font-mono .story-customizer__chip-preview {
-  font-family: "SF Mono", "Fira Code", monospace;
-}
-.story-customizer__chip--font-handwritten .story-customizer__chip-preview {
-  font-family: "Segoe Script", "Comic Sans MS", cursive;
-}
-.story-customizer__chip--font-display .story-customizer__chip-preview {
-  font-family: "Impact", "Arial Black", sans-serif;
-}
-
-/* Swatches */
-.story-customizer__swatches {
-  display: flex;
-  gap: 8px;
-  flex-wrap: wrap;
-}
-
-.story-customizer__swatch {
-  width: 32px;
-  height: 32px;
-  border-radius: 50%;
-  border: 2.5px solid rgba(255, 255, 255, 0.5);
-  cursor: pointer;
-  transition: all 0.2s ease;
-  outline: none;
-  -webkit-tap-highlight-color: transparent;
-  box-shadow: 0 2px 6px rgba(0, 0, 0, 0.12);
-}
-
-.story-customizer__swatch:hover {
-  transform: scale(1.15);
-  box-shadow: 0 3px 10px rgba(0, 0, 0, 0.2);
-}
-
-.story-customizer__swatch--active {
-  border-color: #6b6bf8;
-  box-shadow: 0 0 0 3px rgba(107, 107, 248, 0.35);
-  transform: scale(1.1);
-}
-
-@keyframes fadeSlideUp {
-  from {
-    opacity: 0;
-    transform: translateY(10px);
-  }
-  to {
-    opacity: 1;
-    transform: translateY(0);
-  }
-}
