--- frontend/frontend/src/index.css
@@ -1,47 +1,0 @@
-@import url("https://fonts.googleapis.com/css2?family=DM+Serif+Display&family=Manrope:wght@300;400;500;600;700&display=swap");
-
-:root {
-  font-family: "Manrope", "Segoe UI", sans-serif;
-  line-height: 1.6;
-  font-weight: 400;
-  color: #1d1b22;
-  background-color: #f6f1ea;
-  font-synthesis: none;
-  text-rendering: optimizeLegibility;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-}
-
-* {
-  box-sizing: border-box;
-}
-
-body {
-  margin: 0;
-  min-width: 320px;
-  min-height: 100vh;
-  background:
-    radial-gradient(circle at top left, rgba(255, 204, 156, 0.4), transparent 48%),
-    radial-gradient(circle at 20% 80%, rgba(178, 205, 255, 0.4), transparent 50%),
-    #f6f1ea;
-}
-
-a {
-  color: inherit;
-  text-decoration: none;
-}
-
-h1,
-h2,
-h3 {
-  font-family: "DM Serif Display", "Times New Roman", serif;
-}
-
-button,
-input {
-  font-family: inherit;
-}
-
-#root {
-  min-height: 100vh;
-}
