+++ frontend/src/components/stories/StoryGenerator.css
@@ -1,0 +1,256 @@
+.story-generator {
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  gap: 48px;
+  padding: 2.5rem 6vw 4rem;
+  min-height: calc(100vh - 80px);
+  align-items: start;
+}
+
+
+.story-generator__panel {
+  display: flex;
+  flex-direction: column;
+  gap: 24px;
+  animation: float-in 0.6s ease both;
+}
+
+
+.story-generator__categories {
+  display: flex;
+  gap: 8px;
+  flex-wrap: wrap;
+}
+
+.story-generator__cat-btn {
+  display: inline-flex;
+  align-items: center;
+  gap: 8px;
+  padding: 10px 20px;
+  border-radius: 999px;
+  border: 1.5px solid rgba(29, 27, 34, 0.1);
+  background: rgba(255, 255, 255, 0.6);
+  font-size: 0.88rem;
+  font-weight: 600;
+  color: rgba(29, 27, 34, 0.7);
+  cursor: pointer;
+  transition: background 0.25s ease, color 0.25s ease, border-color 0.25s ease,
+    box-shadow 0.25s ease, transform 0.25s ease;
+  white-space: nowrap;
+  outline: none;
+  -webkit-tap-highlight-color: transparent;
+}
+
+.story-generator__cat-btn span {
+  pointer-events: none;
+}
+
+.story-generator__cat-btn:hover {
+  background: rgba(255, 255, 255, 0.9);
+  border-color: rgba(108, 99, 255, 0.3);
+}
+
+.story-generator__cat-btn.story-generator__cat-btn--active {
+  background: linear-gradient(135deg, #6b6bf8, #b44aff) !important;
+  color: #fff !important;
+  border-color: transparent;
+  box-shadow: 0 4px 16px rgba(107, 107, 248, 0.35);
+  transform: scale(1.02);
+}
+
+.story-generator__cat-btn svg {
+  flex-shrink: 0;
+  width: 16px;
+  height: 16px;
+}
+
+
+.story-generator__caption-wrap {
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+  animation: fadeSlideUp 0.3s ease both;
+}
+
+.story-generator__caption-label {
+  font-size: 0.88rem;
+  font-weight: 600;
+  color: rgba(29, 27, 34, 0.6);
+}
+
+.story-generator__caption {
+  padding: 0.75rem 1rem;
+  border-radius: 14px;
+  border: 1px solid rgba(29, 27, 34, 0.12);
+  font-size: 0.92rem;
+  font-family: inherit;
+  resize: vertical;
+  min-height: 60px;
+  background: rgba(255, 255, 255, 0.7);
+  transition: border-color 0.2s ease, box-shadow 0.2s ease;
+}
+
+.story-generator__caption:focus {
+  outline: none;
+  border-color: rgba(108, 99, 255, 0.5);
+  box-shadow: 0 0 0 3px rgba(108, 99, 255, 0.12);
+}
+
+.story-generator__char-count {
+  align-self: flex-end;
+  font-size: 0.75rem;
+  color: rgba(29, 27, 34, 0.4);
+}
+
+
+.story-generator__preview {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 24px;
+  position: sticky;
+  top: 100px;
+  animation: float-in 0.6s 0.15s ease both;
+}
+
+
+.story-generator__actions {
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+  width: 320px;
+  animation: fadeSlideUp 0.3s ease both;
+}
+
+.story-generator__btn {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 8px;
+  padding: 14px 20px;
+  border-radius: 16px;
+  font-size: 0.95rem;
+  font-weight: 650;
+  cursor: pointer;
+  border: none;
+  transition: transform 0.2s ease, box-shadow 0.2s ease;
+}
+
+.story-generator__btn:not(:disabled):hover {
+  transform: translateY(-2px);
+}
+
+.story-generator__btn:disabled {
+  opacity: 0.6;
+  cursor: not-allowed;
+}
+
+.story-generator__btn--primary {
+  background: linear-gradient(135deg, #22c55e, #059669);
+  color: #fff;
+  box-shadow: 0 8px 24px rgba(34, 197, 94, 0.35);
+}
+
+.story-generator__btn--primary:not(:disabled):hover {
+  box-shadow: 0 12px 32px rgba(34, 197, 94, 0.45);
+}
+
+.story-generator__btn--secondary {
+  background: rgba(255, 255, 255, 0.8);
+  color: #1d1b22;
+  border: 1.5px solid rgba(29, 27, 34, 0.12);
+  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.06);
+}
+
+.story-generator__btn--secondary:hover {
+  background: #fff;
+  border-color: rgba(108, 99, 255, 0.3);
+}
+
+.story-generator__btn-icon {
+  font-size: 1.1rem;
+}
+
+
+.story-generator__result {
+  padding: 10px 16px;
+  border-radius: 12px;
+  font-size: 0.88rem;
+  font-weight: 600;
+  text-align: center;
+  width: 320px;
+  animation: fadeSlideUp 0.25s ease both;
+}
+
+.story-generator__result--success {
+  background: rgba(34, 197, 94, 0.15);
+  color: #15803d;
+}
+
+.story-generator__result--error {
+  background: rgba(239, 68, 68, 0.15);
+  color: #b91c1c;
+}
+
+
+.story-generator__empty {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 12px;
+  padding: 80px 40px;
+  text-align: center;
+  color: rgba(29, 27, 34, 0.35);
+  animation: float-in 0.6s ease both;
+}
+
+.story-generator__empty svg {
+  opacity: 0.4;
+}
+
+.story-generator__empty p {
+  margin: 0;
+  font-size: 1rem;
+}
+
+
+@keyframes fadeSlideUp {
+  from {
+    opacity: 0;
+    transform: translateY(10px);
+  }
+  to {
+    opacity: 1;
+    transform: translateY(0);
+  }
+}
+
+@keyframes float-in {
+  from {
+    opacity: 0;
+    transform: translateY(18px) scale(0.98);
+  }
+  to {
+    opacity: 1;
+    transform: translateY(0) scale(1);
+  }
+}
+
+
+@media (max-width: 900px) {
+  .story-generator {
+    grid-template-columns: 1fr;
+    gap: 32px;
+    padding: 1.5rem 1.5rem 3rem;
+  }
+
+  .story-generator__preview {
+    position: static;
+  }
+
+  .story-generator__actions,
+  .story-generator__result {
+    width: 100%;
+    max-width: 320px;
+  }
+}
