+++ frontend/src/components/stories/index.js
@@ -1,0 +1,4 @@
+export { default as StoryCard } from "./StoryCard";
+export { default as ContentSelector } from "./ContentSelector";
+export { default as StoryCustomizer } from "./StoryCustomizer";
+export { default as StoryGenerator } from "./StoryGenerator";
